requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
drf-yasg==1.20.0
packaging==20.9
Pillow==8.1.2
requests==2.25.1
urllib3==1.26.4
